Initial Code for motor control
This includes motor and servo control with the analog joystick. 

L298N is wired to control both. This may not be the best idea but I'll adjust as I go. 

MEGA2560 is currently driving the device. 

Next, I'll be wiring up both ultrasonic sensors to the front and back along with the lidar camera. I need to discover the best way to visualize the lidar data as it streams in AND store in a database.